+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * [#512 ET Intro Start/End isn't exported](https://github.com/WebTools-NG/WebTools-NG/issues/512)
 * [#514 Bring back EN translations from Romain](https://github.com/WebTools-NG/WebTools-NG/issues/514)
 * [#508 No timestamp in output file](https://github.com/WebTools-NG/WebTools-NG/issues/508)
+* [#507 Download Module](https://github.com/WebTools-NG/WebTools-NG/issues/507)
 
 ## V0.3.17 (20220601)